Added Comments and Additional Method
I have added to the index.js and the general.js test a method for a full site render.

This will return the data from the API.

All tests have completed so far with no issues.
Be sure when testing to have an endpoint that supports HTTPs not only HTTP. Best thing I have used is ngrok to do the HTTPS and HTTP tunneling to the localhost server for testing.